Refactor diagnostics and service information types
Moved AssemblyServiceInformationProvider and IServiceInformationProvider
to Aspire.ResourceServer.Standalone.Server.Diagnostics namespace.
Introduced Constants class for service name and ServerDiagnostics class
for service version. Cleaned up ServiceInformation class by removing
moved components.
