+++ src/server/Diagnostics/AssemblyServiceInformationProvider.cs
@@ -1,0 +1,12 @@
+﻿namespace Aspire.ResourceServer.Standalone.Server.Diagnostics;
+
+internal sealed class AssemblyServiceInformationProvider : IServiceInformationProvider
+{
+    private ServiceInformation ServiceInformation { get; } = new(
+        Constants.ServiceName, ServerDiagnostics.ServiceVersion);
+
+    public ServiceInformation GetServiceInformation()
+    {
+        return ServiceInformation;
+    }
+}
