+++ src/server/Diagnostics/Constants.cs
@@ -1,0 +1,6 @@
+﻿namespace Aspire.ResourceServer.Standalone.Server.Diagnostics;
+
+internal static class Constants
+{
+    public const string ServiceName = "Aspire.ResourceServer.Standalone.Server";
+}
