+++ src/server/Diagnostics/IServiceInformationProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace Aspire.ResourceServer.Standalone.Server.Diagnostics;
+
+public interface IServiceInformationProvider
+{
+    ServiceInformation GetServiceInformation();
+}
